+++ CHANGELOG.md
@@ -59,6 +59,7 @@
 - **`The Eater of Worlds`** has received an audiovisual overhaul, featuring a debris effect system, disgusting movement & digging audio, as well as terrifying attack and pain screeches that make it look like the victim during its battle. Who knew worms could sing?
 - When the last alive boss is killed, a cinematic transition cue is now played in addition to the aforementioned camera focus.
 - Crimson and Corruption enemies now have new hit and death audio.
+- Added sound effects for getting in a bed, as well as sleeping.
 - Leaves particles will now be baked as decals whenever they hit the ground.
 - Introduced many new randomly played ambience sounds (previously we've only had loops).
 ### Critical Strikes